requirements.txt
fastapi[all]
openai
python-dotenv
pydantic
langchain
bs4
tiktoken